+++ php/tests/Application/Command/LocalizeTypeTest.php
@@ -16,7 +16,7 @@ class LocalizeTypeTest extends IndexedTest
         $command = new LocalizeType();
         $command->setIndexDatabase($indexDatabase);
 
-        $this->assertEquals($command->localizeType('C', $path, 1), null);
+        $this->assertEquals('\C', $command->localizeType('C', $path, 1));
         $this->assertEquals('A\C', $command->localizeType('A\C', $path, 5));
         $this->assertEquals('B\C', $command->localizeType('B\C', $path, 10));
         $this->assertEquals('B\DateTime', $command->localizeType('B\DateTime', $path, 10));