+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes for developers
 * Changes to the service
+  * `localizeType` will now return the FQCN instead of null if it couldn't localize a type based on use statements. The reasoning behind this is that you want a localized version of the type you pass, if there are no use statements to localize it, the FQCN is the only way to address the type locally.
   * Return types for functions and types for function parameters will now also include a `typeHint` property that is set to the actual type hint that was specified (the type and resolvedType fall back to the docblock if it is present).
 
 ## 0.9.3